database tests & bug fix & small additions
