--- flockmp/errors.py
@@ -1,7 +1,0 @@
-class FlockError(object):
-    """
-    Handles errors inside the Child processes.
-    """
-
-    def encode(self, error):
-        return error
