+++ flockmp/utils/errors.py
@@ -1,0 +1,2 @@
+class FunctionNotPickled(Exception):
+    "Error inside flock"
