+++ test/test_list.py
@@ -1,5 +1,5 @@
 from flockmp.list import ListAsync
-import pytest
+
 
 class TestList(object):
